created new class MenuEntry and put all functions to create it inside
